+++ src/main/java/br/ufsc/tcc/gamifyEngine/dao/UserDao.java
@@ -13,6 +13,6 @@ public interface UserDao extends CrudRepository<User, Long> {
    * Note that this method is not implemented and its working code will be
    * automagically generated from its signature by Spring Data JPA.
    */
-  public User findById(long userId);
+  public User findById(int userId);
 
 }
