+++ src/main/java/br/ufsc/tcc/gamifyEngine/service/UserService.java
@@ -3,7 +3,7 @@ package br.ufsc.tcc.gamifyEngine.service;
 import br.ufsc.tcc.gamifyEngine.model.User;
 
 public interface UserService {
-	public User getUser(long userId);
+	public User getUser(int userId);
 	public Iterable<User> findAllUsers();
 	public User saveUser(User user);
 }
